fix(EDCO-17): fixed PreHeader default story args
