+++ src/stories/PreHeader/light.stories.tsx
@@ -1,7 +1,6 @@
 import { type ComponentMeta, type ComponentStory } from '@storybook/react';
 import { PreHeader } from '../../components/PreHeader';
 import { defaults } from './defaults';
-import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
 
 export default {
   component: PreHeader,
@@ -15,25 +14,3 @@ const Template: ComponentStory<typeof PreHeader> = (args) => (
 );
 
 export const Standard = Template.bind({});
-Standard.args = {
-  leftCtas: {
-    theme: 'light',
-    ctaButtons: [
-      { text: 'Link', variant: 'naked', color: 'inherit', href: '#' },
-    ],
-  },
-  rightCtas: {
-    theme: 'light',
-    ctaButtons: [
-      {
-        text: 'Assistenza',
-        variant: 'naked',
-        color: 'inherit',
-        startIcon: <HelpOutlineIcon />,
-        onClick: () => {
-          alert('Assistenza');
-        },
-      },
-    ],
-  },
-};
